Intranet: Correcció esborrar promoció.
git-svn-id: https://servidor.casadecultura.cat/public/ccg/trunk/intranet@329 6722b150-0e3f-45da-bd85-83856719fb75

+++ apps/intranet/modules/gestio/actions/actions.class.php
@@ -284,8 +284,9 @@ class gestioActions extends sfActions
       $this->FPromocio = new PromocionsForm($OPromocio);
       $this->EDICIO = true;
     elseif($request->getParameter('BDELETE')): //Esborra
-      $this->PROMOCIO = PromocionsPeer::retrieveByPK($request->getParameter('IDP'));
-      $this->PROMOCIO->delete();
+      $RP = $request->getParameter('promocions');
+      $this->PROMOCIO = PromocionsPeer::retrieveByPK($RP['PromocioID']);
+      if($this->PROMOCIO instanceof Promocions): $this->PROMOCIO->delete(); endif; 
     elseif($request->getParameter('accio') == 'CC'):
     	$this->getUser()->addLogAction('inside','gPromocions');
     endif;
